[UTL-149] add user enum
