+++ .vscode/settings.json
@@ -3,7 +3,7 @@
     ".git": true,
     ".husky": true,
     "node_modules": true,
-    "README.md": true,
+    // "README.md": true,
     ".github": true,
     ".idea": true
   },
