+++ README.md
@@ -92,6 +92,7 @@
 - `UserStatusEnum` - пользовательские статусы
 - `SortingEnum` - варианты сортировки
 - `WorkModeEnum` - список режимов, в которых может работать приложение
+- `UserEnum` - типы пользователя
 - `ProcessEnvironmentKeys` - список всех ключей (process.env) во всех сервисах
 
 - `TupleUnion` - возвращает кортеж, в котором нужно описать все ключи по одному разу
