+++ utils/swagger.ts
@@ -1,10 +1,15 @@
+import { readFileSync } from 'node:fs';
+
 import { DocumentBuilder, SwaggerModule } from '@nestjs/swagger';
+import { replace } from 'lodash';
+
+import { WorkModeEnum } from 'types';
 
 import type { INestApplication } from '@nestjs/common';
 
 interface SetupSwagger {
   app: INestApplication;
-  mode: string;
+  mode?: WorkModeEnum;
   swaggerTitle: string;
   swaggerDescription: string;
   swaggerVersion: string;
@@ -19,7 +24,7 @@ export function setupSwagger({
   swaggerVersion,
   needBearerAuth,
 }: SetupSwagger): void {
-  if (mode === 'test') return;
+  if (mode === WorkModeEnum.Test) return;
 
   const config = new DocumentBuilder()
     .setTitle(swaggerTitle)
@@ -37,4 +42,18 @@ export function setupSwagger({
 
   const document = SwaggerModule.createDocument(app, builderConfig);
   SwaggerModule.setup('swagger', app, document);
+}
+
+export function getVersion(filePath: string) {
+  try {
+    const data: { version?: string } = JSON.parse(readFileSync(filePath, 'utf8'));
+    const version = data?.version;
+    if (typeof version !== 'string') throw new Error('Version not found');
+
+    return replace(version, /[\sv]/g, '');
+  } catch (error) {
+    console.log(error);
+
+    return '0.0.0';
+  }
 }
